Fixed broken build and tests
